Move user profile email lists into separate component
